Refactor Classroom component and update year selection store to improve data and error management
